+++ Knil/App/AppDelegate.swift
@@ -14,6 +14,10 @@ import KnilUIKit
 class AppDelegate: UIResponder, UIApplicationDelegate {
 
     var window: UIWindow?
+    
+    private lazy var currentPasteboardString: String? = {
+        return UIPasteboard.general.string
+    }()
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
@@ -44,6 +48,12 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func applicationDidBecomeActive(_ application: UIApplication) {
         // Restart any tasks that were paused (or not yet started) while the application was inactive. If the application was previously in the background, optionally refresh the user interface.
+
+        if let pasteboardString = UIPasteboard.general.string,
+            pasteboardString != currentPasteboardString {
+            NotificationCenter.default.post(name: .UIPasteboardChanged, object: nil)
+            currentPasteboardString = pasteboardString
+        }
     }
 
     func applicationWillTerminate(_ application: UIApplication) {